Moved getBoundingSphere to base node class
